hice ya todo

+++ grails-app/controllers/Plantilla/UsuarioController.groovy
@@ -18,7 +18,7 @@ class UsuarioController {
         def fecha2 = fecha.substring(0,10)
         [user:editarUsuario, fecha:fecha2]
     }
-    def update(int id){
+    def update(long id){
         def editarUsuario = Usuario.findById(id)
         def fecha = editarUsuario.fechaNac.toString()
         def fecha2 = fecha.substring(0,10)
